% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank data}
\format{
A data frame with 1949 rows and 14 variables
}
\description{
Data from an Italian bank on 1,949 customers. The response y is the amount of money made over a year.
The 13 covariates are possible macroservices the customers can sign up for.
}
\references{
Marco Riani, Andrea Cerioli, and Anthony C. Atkinson (2014).
Monitoring robust regression. Electronic Journal of Statistics, Volume 8, 646-677.
}
\keyword{datasets}
