% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E-package.R
\docType{package}
\name{L2E}
\alias{L2E}
\title{L2E}
\description{
The L2E package is an R implementation of a user-friendly 
computational framework for performing a wide variety of
robust structured regression methods via the L2 criterion.
}
\details{
Please refer to the vignette for examples of how to use
this package.
}
