% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2e_regression_sparse_ncv.R
\name{l2e_regression_sparse_ncv}
\alias{l2e_regression_sparse_ncv}
\title{L2E sparse regression with existing penalization methods}
\usage{
l2e_regression_sparse_ncv(
  y,
  X,
  beta,
  tau,
  lambda,
  penalty,
  max_iter = 100,
  tol = 1e-04,
  Show.Time = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{beta}{Initial vector of regression coefficients}

\item{tau}{Initial precision estimate}

\item{lambda}{Tuning parameter}

\item{penalty}{Available penalties include lasso, MCP and SCAD.}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}

\item{Show.Time}{Report the computing time}
}
\description{
\code{l2e_regression_sparse_ncv} performs robust sparse regression under the L2 criterion. Available penalties include lasso, MCP and SCAD.
}
