% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2e_regression_TF_lasso.R
\name{l2e_regression_TF_lasso}
\alias{l2e_regression_TF_lasso}
\title{L2E trend filtering regression with Lasso penalization}
\usage{
l2e_regression_TF_lasso(
  y,
  X,
  beta,
  tau,
  D,
  lambda = 1,
  max_iter = 100,
  tol = 1e-04,
  Show.Time = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{beta}{Initial vector of regression coefficients}

\item{tau}{Initial precision estimate}

\item{D}{The fusion matrix}

\item{lambda}{The tuning parameter}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}

\item{Show.Time}{Report the computing time}
}
\description{
\code{l2e_regression_TF_lasso} performs robust trend filtering regression under the L2 criterion with Lasso penalty
}
