% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGetDkn.R
\name{myGetDkn}
\alias{myGetDkn}
\title{Compute kth order differencing matrix}
\usage{
myGetDkn(k, n)
}
\arguments{
\item{k}{Order of the differencing matrix}

\item{n}{Number of time points}
}
\value{
Returns a Matrix object as the kth order differencing matrix
}
\description{
\code{myGetDkn} computes the kth order differencing matrix for use in
trend filtering regression
}
