% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{objective}
\alias{objective}
\title{Objective function of the L2E regression - eta}
\usage{
objective(eta, r, method = "mean")
}
\arguments{
\item{eta}{The current estimate of eta}

\item{r}{Vector of residuals}

\item{method}{Mean or median}
}
\value{
Returns the output of the objective function (scalar)
}
\description{
\code{objective} computes the objective of the L2E regression in terms of eta
}
