% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{objective_tau}
\alias{objective_tau}
\title{Objective function of the L2E regression - tau}
\usage{
objective_tau(tau, r, method = "mean")
}
\arguments{
\item{tau}{The current estimate of tau}

\item{r}{Vector of residuals}

\item{method}{Mean or median}
}
\value{
Returns the output of the objective function (scalar)
}
\description{
\code{objective_tau} computes the objective of the L2E regression in terms of tau
}
