% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta_isotonic.R
\name{update_beta_isotonic}
\alias{update_beta_isotonic}
\title{Beta update in LTE isotonic regression}
\usage{
update_beta_isotonic(y, b, tau, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{y}{Response vector}

\item{b}{Current estimate for beta}

\item{tau}{Current estimate for tau}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\description{
\code{update_beta_isotonic} Function for updating beta in LTE isotonic regression
}
