% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta_MM_isotonic.R
\name{update_beta_MM_isotonic}
\alias{update_beta_MM_isotonic}
\title{Beta update in L2E isotonic regression - MM}
\usage{
update_beta_MM_isotonic(y, beta, tau, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{y}{response}

\item{beta}{initial vector of regression coefficients}

\item{tau}{precision estimate}

\item{max_iter}{maximum number of iterations}

\item{tol}{relative tolerance}
}
\value{
Returns a list object containing the new estimate for beta (vector) and the number of iterations (scalar) the update step utilized
}
\description{
\code{update_beta_MM_isotonic} updates beta for L2E isotonic regression using MM
}
