% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta_qr.R
\name{update_beta_qr}
\alias{update_beta_qr}
\title{Beta update in LTE multivariate regression}
\usage{
update_beta_qr(y, X, QRF, tau, b, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{QRF}{QR factorization object for X (obtained via `QRF=qr(X)`)}

\item{tau}{Current estimate for tau}

\item{b}{Current estimate for beta}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the new estimate for beta (vector) and the number of iterations (scalar) the update step utilized
}
\description{
\code{update_beta_qr} Function for updating beta in LTE multivariate regression via a QR solve
}
