% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_eta_bktk.R
\name{update_eta_bktk}
\alias{update_eta_bktk}
\title{Eta update using Newton's method with backtracking}
\usage{
update_eta_bktk(r, eta, max_iter = 100, tol = 1e-10)
}
\arguments{
\item{r}{Vector of residual}

\item{eta}{Initial estimate of eta}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the new estimate for eta (scalar),
the number of iterations (scalar) the update step utilized,
the eta and objective function solution paths (vectors), and
the first and second derivatives calculated via Newton's method (vectors)
}
\description{
\code{update_eta_bktk} updates the precision parameter tau = e^eta for L2E regression using Newton's method
}
