% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau_R.R
\name{update_tau_R}
\alias{update_tau_R}
\title{Tau update function}
\usage{
update_tau_R(r, tau, sd_y, max_iter = 100, tol = 1e-10)
}
\arguments{
\item{r}{Residual vector}

\item{tau}{Current estimate for tau}

\item{sd_y}{Standard deviation of y}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the new estimate for tau (scalar) and the number of iterations (scalar) the update step utilized
}
\description{
\code{update_tau_R} Function for updating tau
}
