% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta_convex.R
\name{update_beta_convex}
\alias{update_beta_convex}
\title{Beta update in LTE convex regression}
\usage{
update_beta_convex(y, b, tau, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{y}{Response vector}

\item{b}{Current estimate for beta}

\item{tau}{Current estimate for tau}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the new estimate for beta (vector) and the number of iterations (scalar) the update step utilized
}
\description{
\code{update_beta_convex} Function for updating beta in LTE convex regression
}
