% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta_isotonic.R
\name{update_beta_isotonic}
\alias{update_beta_isotonic}
\title{Beta update in L2E isotonic regression - PG}
\usage{
update_beta_isotonic(y, b, tau, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{y}{Response vector}

\item{b}{Current estimate for beta}

\item{tau}{Current estimate for tau}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the new estimate for beta (vector) and the number of iterations (scalar) the update step utilized
}
\description{
\code{update_beta_isotonic} updates beta for L2E isotonic regression using PG
}
